refactor litpose worker
